Threat Carver/requirements.txt
streamlit>=1.10.0
pandas>=1.3.0
plotly>=5.3.0
requests>=2.26.0
pyyaml>=6.0
